Move code from minecode/collectors to original location #49 #285
