+++ minecode/management/commands/get_maven_release_dates.py
@@ -15,8 +15,8 @@ import sys
 import requests
 
 from minecode.management.commands import VerboseCommand
-from minecode.collectors.maven import collect_links_from_text
-from minecode.collectors.maven import filter_for_artifacts
+from minecode.visitors.maven import collect_links_from_text
+from minecode.visitors.maven import filter_for_artifacts
 from packagedb.models import Package
