+++ minecode/management/commands/maven_crawler.py
@@ -10,7 +10,7 @@
 import logging
 import sys
 
-from minecode.collectors.maven import crawl_maven_repo_from_root
+from minecode.visitors.maven import crawl_maven_repo_from_root
 from minecode.management.commands import VerboseCommand
