+++ minecode/tests/test_maven.py
@@ -17,7 +17,6 @@ from unittest import mock
 
 from django.test import TestCase as DjangoTestCase
 
-from minecode.collectors import maven as maven_collector
 from minecode.management.commands.run_map import map_uri
 from minecode.management.commands.run_visit import visit_uri
 from minecode.mappers import maven as maven_mapper
@@ -701,7 +700,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         )
 
     def test_get_pom_text(self, regen=False):
-        pom_contents = maven_collector.get_pom_text(
+        pom_contents = maven_visitor.get_pom_text(
             namespace=self.scan_package.namespace,
             name=self.scan_package.name,
             version=self.scan_package.version
@@ -712,7 +711,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         self.assertEqual(self.expected_pom_contents, pom_contents)
 
     def test_get_package_sha1(self):
-        sha1 = maven_collector.get_package_sha1(self.scan_package)
+        sha1 = maven_visitor.get_package_sha1(self.scan_package)
         expected_sha1 = '60c708f55deeb7c5dfce8a7886ef09cbc1388eca'
         self.assertEqual(expected_sha1, sha1)
 
@@ -720,7 +719,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         package_count = packagedb.models.Package.objects.all().count()
         self.assertEqual(0, package_count)
         package_url = PackageURL.from_string(self.scan_package.purl)
-        maven_collector.map_maven_package(package_url, packagedb.models.PackageContentType.BINARY)
+        maven_visitor.map_maven_package(package_url, packagedb.models.PackageContentType.BINARY)
         package_count = packagedb.models.Package.objects.all().count()
         self.assertEqual(1, package_count)
         package = packagedb.models.Package.objects.all().first()
@@ -732,7 +731,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         self.assertEqual(0, package_count)
         custom_repo_purl = "pkg:maven/org.eclipse.core/runtime@20070801?repository_url=https://packages.atlassian.com/mvn/maven-atlassian-external/"
         package_url = PackageURL.from_string(custom_repo_purl)
-        maven_collector.map_maven_package(package_url, packagedb.models.PackageContentType.BINARY)
+        maven_visitor.map_maven_package(package_url, packagedb.models.PackageContentType.BINARY)
         package_count = packagedb.models.Package.objects.all().count()
         self.assertEqual(1, package_count)
         package = packagedb.models.Package.objects.all().first()
@@ -747,7 +746,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         sources_download_url = 'https://repo1.maven.org/maven2/org/apache/twill/twill-core/0.12.0/twill-core-0.12.0-sources.jar'
         package_count = packagedb.models.Package.objects.all().count()
         self.assertEqual(0, package_count)
-        maven_collector.process_request(purl_str)
+        maven_visitor.process_request(purl_str)
         package_count = packagedb.models.Package.objects.all().count()
         self.assertEqual(2, package_count)
         purls = [
@@ -765,7 +764,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         pom_loc = self.get_test_loc('maven/pom/ant-antlr-1.10.1.pom')
         with open(pom_loc) as f:
             pom_text = f.read()
-        parent_pom_text = maven_collector.fetch_parent(pom_text)
+        parent_pom_text = maven_visitor.fetch_parent(pom_text)
         expected_loc = self.get_test_loc('maven/pom/ant-parent-1.10.1.pom')
 
         if regen:
@@ -780,7 +779,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
         pom_loc = self.get_test_loc('maven/pom/pulsar-client-1x-2.5.1.pom')
         with open(pom_loc) as f:
             pom_text = f.read()
-        ancestor_pom_texts = list(maven_collector.get_ancestry(pom_text))
+        ancestor_pom_texts = list(maven_visitor.get_ancestry(pom_text))
         expected_ancestor_pom_texts = []
         for expected_loc in [
             self.get_test_loc('maven/pom/apache-18.pom'),
@@ -814,7 +813,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
             'Java',
             text=parent_pom_text
         )
-        package = maven_collector.merge_parent(package, parent_package)
+        package = maven_visitor.merge_parent(package, parent_package)
         expected_after_loc = self.get_test_loc('maven/pom/ant-antlr-1.10.1-package_after.json')
         self.check_expected_results(package.to_dict(), expected_after_loc, regen=regen)
 
@@ -841,15 +840,15 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
                 pom_text = f.read()
             ancestor_pom_texts.append(pom_text)
 
-        maven_collector.merge_ancestors(ancestor_pom_texts, package)
+        maven_visitor.merge_ancestors(ancestor_pom_texts, package)
         expected_after_loc = self.get_test_loc('maven/pom/pulsar-client-1x-2.5.1-package_after.json')
         self.check_expected_results(package.to_dict(), expected_after_loc, regen=regen)
 
-    @mock.patch("minecode.collectors.maven.get_pom_text")
+    @mock.patch("minecode.visitors.maven.get_pom_text")
     def test_get_merged_ancestor_package_from_maven_package(self, get_pom_text_mock, regen=False):
         get_pom_text_mock.return_value = ""
         ancestor_pom_texts = []
-        with patch("minecode.collectors.maven.get_ancestry") as mock_get_ancestry:
+        with patch("minecode.visitors.maven.get_ancestry") as mock_get_ancestry:
             for loc in [
                 self.get_test_loc('maven/pom/apache-18.pom'),
                 self.get_test_loc('maven/pom/pulsar-2.5.1.pom'),
@@ -866,7 +865,7 @@ class MavenPriorityQueueTests(JsonBasedTesting, DjangoTestCase):
                 type="maven",
                 download_url="https://repo1.maven.org/maven2/org/apache/pulsar/pulsar-client/2.5.1/pulsar-client-2.5.1.jar",
             )
-            merged_package = maven_collector.get_merged_ancestor_package_from_maven_package(package=db_package)
+            merged_package = maven_visitor.get_merged_ancestor_package_from_maven_package(package=db_package)
             expected_loc = self.get_test_loc('maven/pom/pulsar-client-merged-ancestor-package.json')
             self.check_expected_results(merged_package.to_dict(), expected_loc, regen=regen)
 
@@ -877,60 +876,60 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
     def test_check_if_file_name_is_linked_on_page(self):
         links = ['foo/', 'bar/', 'baz/']
         self.assertTrue(
-            maven_collector.check_if_file_name_is_linked_on_page('foo/', links)
+            maven_visitor.check_if_file_name_is_linked_on_page('foo/', links)
         )
         self.assertFalse(
-            maven_collector.check_if_file_name_is_linked_on_page('qux/', links)
+            maven_visitor.check_if_file_name_is_linked_on_page('qux/', links)
         )
 
     def test_check_if_page_has_pom_files(self):
         links1 = ['foo/', 'bar.jar', 'bar.pom']
         links2 = ['foo/', 'bar.jar']
-        self.assertTrue(maven_collector.check_if_page_has_pom_files(links1))
-        self.assertFalse(maven_collector.check_if_page_has_pom_files(links2))
+        self.assertTrue(maven_visitor.check_if_page_has_pom_files(links1))
+        self.assertFalse(maven_visitor.check_if_page_has_pom_files(links2))
 
     def test_check_if_page_has_directories(self):
         links1 = ['foo/', 'bar/', 'baz/']
         links2 = ['../', 'bar.pom', 'bar.jar']
-        self.assertTrue(maven_collector.check_if_page_has_directories(links1))
-        self.assertFalse(maven_collector.check_if_page_has_directories(links2))
+        self.assertTrue(maven_visitor.check_if_page_has_directories(links1))
+        self.assertFalse(maven_visitor.check_if_page_has_directories(links2))
 
     def test_check_if_package_version_page(self):
         links1 = ['../', 'bar.pom', 'bar.jar']
         links2 = ['../', 'foo/', 'bar/', 'baz/']
-        self.assertTrue(maven_collector.check_if_package_version_page(links1))
-        self.assertFalse(maven_collector.check_if_package_version_page(links2))
+        self.assertTrue(maven_visitor.check_if_package_version_page(links1))
+        self.assertFalse(maven_visitor.check_if_package_version_page(links2))
 
     def test_check_if_package_page(self):
         links1 = ['../', 'maven-metadata.xml']
         links2 = ['../', 'bar.pom', 'bar.jar']
-        self.assertTrue(maven_collector.check_if_package_page(links1))
-        self.assertFalse(maven_collector.check_if_package_page(links2))
+        self.assertTrue(maven_visitor.check_if_package_page(links1))
+        self.assertFalse(maven_visitor.check_if_package_page(links2))
 
     def test_check_if_maven_root(self):
         links1 = ['../', 'archetype-catalog.xml']
         links2 = ['../', 'bar.pom', 'bar.jar']
-        self.assertTrue(maven_collector.check_if_maven_root(links1))
-        self.assertFalse(maven_collector.check_if_maven_root(links2))
+        self.assertTrue(maven_visitor.check_if_maven_root(links1))
+        self.assertFalse(maven_visitor.check_if_maven_root(links2))
 
     @mock.patch('requests.get')
     def test_check_on_page(self, mock_request_get):
-        checker = maven_collector.check_if_page_has_pom_files
+        checker = maven_visitor.check_if_page_has_pom_files
         mock_request_get.return_value.ok = True
         mock_request_get.return_value.text = '<a href="parent-7.11.0.pom" title="parent-7.11.0.pom">parent-7.11.0.pom</a>'
-        self.assertTrue(maven_collector.check_on_page('https://repo1.maven.org/maven2/net/shibboleth/parent/7.11.0/', checker))
+        self.assertTrue(maven_visitor.check_on_page('https://repo1.maven.org/maven2/net/shibboleth/parent/7.11.0/', checker))
 
     @mock.patch('requests.get')
     def test_is_maven_root(self, mock_request_get):
         mock_request_get.return_value.ok = True
         mock_request_get.return_value.text = '<a href="archetype-catalog.xml" title="archetype-catalog.xml">archetype-catalog.xml</a>'
-        self.assertTrue(maven_collector.is_maven_root('https://repo1.maven.org/maven2/'))
+        self.assertTrue(maven_visitor.is_maven_root('https://repo1.maven.org/maven2/'))
 
     @mock.patch('requests.get')
     def test_is_package_page(self, mock_request_get):
         mock_request_get.return_value.ok = True
         mock_request_get.return_value.text = '<a href="maven-metadata.xml" title="maven-metadata.xml">maven-metadata.xml</a>'
-        self.assertTrue(maven_collector.is_package_page('https://repo1.maven.org/maven2/xml-apis/xml-apis/'))
+        self.assertTrue(maven_visitor.is_package_page('https://repo1.maven.org/maven2/xml-apis/xml-apis/'))
 
     @mock.patch('requests.get')
     def test_is_package_version_page(self, mock_request_get):
@@ -939,11 +938,11 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         <a href="../" title="../">../</a>
         <a href="parent-7.11.0.pom" title="parent-7.11.0.pom">parent-7.11.0.pom</a>
         '''
-        self.assertTrue(maven_collector.is_package_version_page('https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/'))
+        self.assertTrue(maven_visitor.is_package_version_page('https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/'))
 
     def test_url_parts(self):
         url = 'https://example.com/foo/bar/baz.jar'
-        scheme, netloc, path_segments = maven_collector.url_parts(url)
+        scheme, netloc, path_segments = maven_visitor.url_parts(url)
         self.assertEqual('https', scheme)
         self.assertEqual('example.com', netloc)
         self.assertEquals(['foo', 'bar', 'baz.jar'], path_segments)
@@ -955,7 +954,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         url = 'https://example.com/foo/bar/baz.jar'
         self.assertEqual(
             url,
-            maven_collector.create_url(scheme, netloc, path_segments)
+            maven_visitor.create_url(scheme, netloc, path_segments)
         )
 
     @mock.patch('requests.get')
@@ -964,7 +963,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         mock_request_get.return_value.text = '<a href="archetype-catalog.xml" title="archetype-catalog.xml">archetype-catalog.xml</a>'
         self.assertEqual(
             'https://repo1.maven.org/maven2',
-            maven_collector.get_maven_root('https://repo1.maven.org/maven2/net/shibboleth/parent/7.11.0/')
+            maven_visitor.get_maven_root('https://repo1.maven.org/maven2/net/shibboleth/parent/7.11.0/')
         )
 
     @mock.patch('requests.get')
@@ -994,7 +993,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
             package_version_page
         ]
 
-        namespace, package_name, package_version = maven_collector.determine_namespace_name_version_from_url(url, root_url)
+        namespace, package_name, package_version = maven_visitor.determine_namespace_name_version_from_url(url, root_url)
         self.assertEqual('xml-apis', namespace)
         self.assertEqual('xml-apis', package_name)
         self.assertEqual('1.0.b2', package_version)
@@ -1030,7 +1029,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         ]
 
         self.assertEqual(0, ImportableURI.objects.all().count())
-        maven_collector.add_to_import_queue(url, root_url )
+        maven_visitor.add_to_import_queue(url, root_url )
         self.assertEqual(1, ImportableURI.objects.all().count())
         importable_uri = ImportableURI.objects.get(uri=url)
         self.assertEqual('pkg:maven/xml-apis/xml-apis', importable_uri.package_url)
@@ -1046,7 +1045,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         }
         self.assertEqual(
             expected,
-            maven_collector.filter_only_directories(timestamps_by_links)
+            maven_visitor.filter_only_directories(timestamps_by_links)
         )
 
     def test_filter_for_artifacts(self):
@@ -1084,10 +1083,10 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
             'foo.xar': '2023-09-28',
             'foo.zip': '2023-09-28',
         }
-        self.assertEqual(expected, maven_collector.filter_for_artifacts(timestamps_by_links))
+        self.assertEqual(expected, maven_visitor.filter_for_artifacts(timestamps_by_links))
 
     def test_collect_links_from_text(self):
-        filter = maven_collector.filter_only_directories
+        filter = maven_visitor.filter_only_directories
         text = '''
         <a href="../">../</a>
         <a href="1.0.b2/" title="1.0.b2/">1.0.b2/</a>
@@ -1101,11 +1100,11 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         }
         self.assertEqual(
             expected,
-            maven_collector.collect_links_from_text(text, filter=filter)
+            maven_visitor.collect_links_from_text(text, filter=filter)
         )
 
     def test_create_absolute_urls_for_links(self):
-        filter = maven_collector.filter_only_directories
+        filter = maven_visitor.filter_only_directories
         text = '''
         <a href="../">../</a>
         <a href="1.0.b2/" title="1.0.b2/">1.0.b2/</a>
@@ -1120,7 +1119,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         }
         self.assertEqual(
             expected,
-            maven_collector.create_absolute_urls_for_links(text, url, filter=filter)
+            maven_visitor.create_absolute_urls_for_links(text, url, filter=filter)
         )
 
     @mock.patch('requests.get')
@@ -1138,7 +1137,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
             'https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/': '2005-09-20 05:53',
             'https://repo1.maven.org/maven2/xml-apis/xml-apis/1.2.01/': '2010-02-03 21:05'
         }
-        self.assertEqual(expected, maven_collector.get_directory_links(url))
+        self.assertEqual(expected, maven_visitor.get_directory_links(url))
 
     @mock.patch('requests.get')
     def test_get_artifact_links(self, mock_request_get):
@@ -1154,7 +1153,7 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         expected = {
             'https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/xml-apis-1.0.b2.jar': '2005-09-20 05:53',
         }
-        self.assertEqual(expected, maven_collector.get_artifact_links(url))
+        self.assertEqual(expected, maven_visitor.get_artifact_links(url))
 
     def test_crawl_to_package(self):
         pass
@@ -1167,14 +1166,14 @@ class MavenCrawlerFunctionsTest(JsonBasedTesting, DjangoTestCase):
         sha1 = '3136ca936f64c9d68529f048c2618bd356bf85c9'
         mock_request_get.return_value.ok = True
         mock_request_get.return_value.text = sha1
-        self.assertEqual(sha1, maven_collector.get_artifact_sha1('https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/xml-apis-1.0.b2.jar.sha1'))
+        self.assertEqual(sha1, maven_visitor.get_artifact_sha1('https://repo1.maven.org/maven2/xml-apis/xml-apis/1.0.b2/xml-apis-1.0.b2.jar.sha1'))
 
     def test_get_classifier_from_artifact_url(self):
         artifact_url = 'https://repo1.maven.org/maven2/net/alchim31/livereload-jvm/0.2.0/livereload-jvm-0.2.0-onejar.jar'
         package_version_page_url = 'https://repo1.maven.org/maven2/net/alchim31/livereload-jvm/0.2.0/'
         package_name = 'livereload-jvm'
         package_version = '0.2.0'
-        classifier = maven_collector.get_classifier_from_artifact_url(
+        classifier = maven_visitor.get_classifier_from_artifact_url(
             artifact_url,
             package_version_page_url,
             package_name,
