+++ packagedb/find_source_repo.py
@@ -21,7 +21,7 @@ from packageurl.contrib.purl2url import purl2url
 from scancode.api import get_urls as get_urls_from_location
 
 from minecode.model_utils import add_package_to_scan_queue
-from minecode.collectors.maven import get_merged_ancestor_package_from_maven_package
+from minecode.visitors.maven import get_merged_ancestor_package_from_maven_package
 from packagedb.models import Package
 from packagedb.models import PackageContentType
 from packagedb.models import PackageSet
