+++ packagedb/tests/testfiles/api/resource-filter_by_checksums-expected.json
@@ -11,9 +11,9 @@
     "sha256":"testsha2561",
     "sha512":"testsha5121",
     "git_sha1":"testgit_sha11",
-    "mime_type":"",
-    "file_type":"",
-    "programming_language":"",
+    "mime_type":null,
+    "file_type":null,
+    "programming_language":null,
     "is_binary":false,
     "is_text":false,
     "is_archive":false,
@@ -44,9 +44,9 @@
     "sha256":"testsha2562",
     "sha512":"testsha5122",
     "git_sha1":"testgit_sha12",
-    "mime_type":"",
-    "file_type":"",
-    "programming_language":"",
+    "mime_type":null,
+    "file_type":null,
+    "programming_language":null,
     "is_binary":false,
     "is_text":false,
     "is_archive":false,
